split code to different components
